add yes no popup

+++ Prefabs/FileItem/file_item.gd
@@ -7,7 +7,8 @@ func initialize_string(_text: String) -> void:
 	self.custom_minimum_size = Vector2(1080, 150)
 
 func _on_delete_button_pressed() -> void:
-	AppManager.delete_list_file(item_label.text)
+	if await AppManager.create_yes_no_input("Are you sure?"):
+		AppManager.delete_list_file(item_label.text)
 
 func _on_file_button_pressed() -> void:
 	AppManager.working_file_name = item_label.text
